Change entrypoints to scripts
Change entrypoints to scripts that wrap starting up sshd in the background.
Sshd must be started as root or it cannot access it's config files.
The wrapper runs vllm or tgis from the su command as the vllm user. Since
we're root, it doesn't require a password to change uids.